Change the ON button led configuration
Now LED turns off when command is accept.

+++ src/Devices/Thermocouple.cpp
@@ -5,7 +5,7 @@ void Thermocouple::setup() {
 }
 
 uint16_t Thermocouple::read() {
-    uint16_t temperature = analogRead(THERMOCOUPLE);
+    unsigned int temperature = analogRead(THERMOCOUPLE);
     return temperature;
 }
